Separate classes in http package into subpackages
The org.edx.mobile.http package was getting crowded, with many different
types of HTTP-related classes residing at the top level. This commit
separates them into logical subpackages for better organization.
Independent packages have been created for interceptors, authenticators,
utility classes, and custom callback classes. This mostly clears up the
base package except for a few old and deprecated classes dealing with
the (deprecated) Apache HTTP client, which should be removed soon as
well. There was also an unused OutboundUrlSpan class that has also been
removed along with this cleanup.
