+++ OpenEdXMobile/src/main/java/org/edx/mobile/http/callback/CallTrigger.java
@@ -1,4 +1,4 @@
-package org.edx.mobile.http;
+package org.edx.mobile.http.callback;
 
 import android.support.annotation.NonNull;
