+++ OpenEdXMobile/src/main/java/org/edx/mobile/http/callback/ErrorHandlingOkCallback.java
@@ -1,32 +1,30 @@
-package org.edx.mobile.http;
+package org.edx.mobile.http.callback;
 
 import android.content.Context;
 import android.support.annotation.NonNull;
 import android.support.annotation.Nullable;
 
+import org.edx.mobile.http.HttpResponseStatusException;
 import org.edx.mobile.util.images.ErrorUtils;
 import org.edx.mobile.view.common.TaskMessageCallback;
 import org.edx.mobile.view.common.TaskProgressCallback;
 
 import java.io.IOException;
 
-import retrofit2.Call;
-import retrofit2.Callback;
-import retrofit2.Response;
+import okhttp3.Call;
+import okhttp3.Callback;
+import okhttp3.Response;
 import roboguice.RoboGuice;
 
 /**
- * Generic abstract implementation of Retrofit's
- * {@link retrofit2.Callback} interface, that takes care of delivering
- * status and error information to the proper callbacks. It also
- * provides (and delegates to) a simpler callback interface for
- * subclasses, stripping out unnecessary parameters, and redirecting
- * all responses with error codes to the failure callback method (as it
- * used to be in the implementation in Retrofit 1).
- *
- * @param <T> The successful response body type.
+ * Generic abstract implementation of OkHttps's {@link Callback}
+ * interface, that takes care of delivering status and error information
+ * to the proper callbacks. It also provides (and delegates to) a simpler
+ * callback interface for subclasses, stripping out unnecessary
+ * parameters, and redirecting all responses with error codes to the
+ * failure callback method.
  */
-public abstract class ErrorHandlingCallback<T> implements Callback<T> {
+public abstract class ErrorHandlingOkCallback implements Callback {
     /**
      * A Context for resolving the error message strings.
      */
@@ -65,8 +63,8 @@ public abstract class ErrorHandlingCallback<T> implements Callback<T> {
      * @param callTrigger The trigger for initiating the call. This is used to determine the type of
      *                    error message to deliver.
      */
-    public ErrorHandlingCallback(@NonNull final Context context,
-                                 @NonNull final CallTrigger callTrigger) {
+    public ErrorHandlingOkCallback(@NonNull final Context context,
+                                   @NonNull final CallTrigger callTrigger) {
         this(context, callTrigger,
                 context instanceof TaskProgressCallback ? (TaskProgressCallback) context : null,
                 context instanceof TaskMessageCallback ? (TaskMessageCallback) context : null);
@@ -87,9 +85,9 @@ public abstract class ErrorHandlingCallback<T> implements Callback<T> {
      *                         initiation, it assumes that it's being initiated immediately, and
      *                         thus invokes that start callback immediately as well.
      */
-    public ErrorHandlingCallback(@NonNull final Context context,
-                                 @NonNull final CallTrigger callTrigger,
-                                 @Nullable final TaskProgressCallback progressCallback) {
+    public ErrorHandlingOkCallback(@NonNull final Context context,
+                                   @NonNull final CallTrigger callTrigger,
+                                   @Nullable final TaskProgressCallback progressCallback) {
         this(context, callTrigger,
                 progressCallback,
                 context instanceof TaskMessageCallback ? (TaskMessageCallback) context : null);
@@ -107,9 +105,9 @@ public abstract class ErrorHandlingCallback<T> implements Callback<T> {
      *                    error message to deliver.
      * @param messageCallback The callback to invoke for delivering any error messages.
      */
-    public ErrorHandlingCallback(@NonNull final Context context,
-                                 @NonNull final CallTrigger callTrigger,
-                                 @Nullable final TaskMessageCallback messageCallback) {
+    public ErrorHandlingOkCallback(@NonNull final Context context,
+                                   @NonNull final CallTrigger callTrigger,
+                                   @Nullable final TaskMessageCallback messageCallback) {
         this(context, callTrigger,
                 context instanceof TaskProgressCallback ? (TaskProgressCallback) context : null,
                 messageCallback);
@@ -127,10 +125,10 @@ public abstract class ErrorHandlingCallback<T> implements Callback<T> {
      *                         thus invokes that start callback immediately as well.
      * @param messageCallback The callback to invoke for delivering any error messages.
      */
-    public ErrorHandlingCallback(@NonNull final Context context,
-                                 @NonNull final CallTrigger callTrigger,
-                                 @Nullable final TaskProgressCallback progressCallback,
-                                 @Nullable final TaskMessageCallback messageCallback) {
+    public ErrorHandlingOkCallback(@NonNull final Context context,
+                                   @NonNull final CallTrigger callTrigger,
+                                   @Nullable final TaskProgressCallback progressCallback,
+                                   @Nullable final TaskMessageCallback messageCallback) {
         this.context = context;
         this.callTrigger = callTrigger;
         this.progressCallback = progressCallback;
@@ -143,59 +141,64 @@ public abstract class ErrorHandlingCallback<T> implements Callback<T> {
     }
 
     /**
-     * The original callback method invoked by Retrofit upon receiving an HTTP response. This method
+     * The original callback method invoked by OkHttp upon receiving an HTTP response. This method
      * definition provides extra information that's not needed by most individual callback
      * implementations, and is also invoked when HTTP error status codes are encountered (forcing
      * the implementation to manually check for success in each case). Therefore this implementation
-     * delegates to {@link #onResponse(Object)} in the case where it receives a successful HTTP
-     * status code, and to {@link #onFailure(Call, Throwable)} otherwise, passing an instance of
+     * delegates to {@link #onResponse(Response)} in the case where it receives a successful HTTP
+     * status code, and to {@link #onFailure(Throwable)} otherwise, passing an instance of
      * {@link HttpResponseStatusException} with the relevant error status code. This method is
      * declared as final, as subclasses are meant to be implementing the abstract
-     * {@link #onResponse(Object)} method instead of this one.
+     * {@link #onResponse(Response)} method instead of this one.
      * <p>
-     * This implementation takes care of invoking the callback for request process completion.
+     * This implementation takes care of delivering the appropriate error message to it's registered
+     * callback, and invoking the callback for request process completion.
      *
      * @param call The Call object that was used to enqueue the request.
      * @param response The HTTP response data.
      */
     @Override
-    public final void onResponse(@NonNull Call<T> call, @NonNull Response<T> response) {
+    public final void onResponse(@NonNull Call call, @NonNull Response response) {
         if (!response.isSuccessful()) {
-            onFailure(call, new HttpResponseStatusException(response));
+            deliverFailure(new HttpResponseStatusException(response));
         } else {
             if (progressCallback != null) {
                 progressCallback.finishProcess();
             }
-            onResponse(response.body());
+            onResponse(response);
         }
     }
 
     /**
-     * The original callback method invoked by Retrofit upon failure to receive an HTTP response,
-     * whether due to encountering a network error while waiting for the response, or some other
-     * unexpected error while constructing the request or processing the response. It's also invoked
-     * by the {@link #onResponse(Call, Response)} implementation when it receives an HTTP error
-     * status code. However, this method definition provides extra information that's not needed by
-     * most individual callback implementations, so this implementation only delegates to
-     * {@link #onFailure(Throwable)}.
-     * <p>
-     * This implementation takes care of delivering the appropriate error message to it's registered
-     * callback, and invoking the callback for request process completion. It should only be
-     * overridden if the subclass wants to handle or control these actions itself; otherwise
-     * subclasses should override the empty {@link #onFailure(Throwable)} callback method instead.
+     * The original callback method invoked by OkHttp upon failure to receive an HTTP response,
+     * whether due to cancellation, a connectivity problem, or a timeout. This method definition
+     * provides extra information that's not needed by most individual callback implementations, so
+     * this implementation only delegates to {@link #onFailure(Throwable)}. This method is declared
+     * as final, as subclasses are meant to be implementing the abstract
+     * {@link #onResponse(Response)} method instead of this one.
      *
      * @param call The Call object that was used to enqueue the request.
-     * @param error An {@link IOException} if the request failed due to a network failure, an
-     *              {HttpResponseStatusException} if the failure was due to receiving an error code,
-     *              or any {@link Throwable} implementation if one was thrown unexpectedly while
-     *              creating the request or processing the response.
+     * @param error The cause of the request being interrupted.
      */
     @Override
-    public void onFailure(@NonNull Call<T> call, @NonNull Throwable error) {
+    public final void onFailure(@NonNull Call call, @NonNull IOException error) {
+        deliverFailure(error);
+    }
+
+    /**
+     * Convenience method for taking care of invoking the failure callbacks and invoking the new
+     * failure callback method in the case on an error, used by both the original failure callback
+     * method, and the success callback method in the case where it encounters an HTTP error status
+     * code.
+     *
+     * @param error An {@link IOException} if the request failed due to a network failure, or an
+     *              {HttpResponseStatusException} if the failure was due to receiving an error code.
+     */
+    private void deliverFailure(@NonNull final Throwable error) {
         if (progressCallback != null) {
             progressCallback.finishProcess();
         }
-        if (messageCallback != null && !call.isCanceled()) {
+        if (messageCallback != null) {
             messageCallback.onMessage(callTrigger.getMessageType(),
                     ErrorUtils.getErrorMessage(error, context));
         }
@@ -205,20 +208,16 @@ public abstract class ErrorHandlingCallback<T> implements Callback<T> {
     /**
      * Callback method for a successful HTTP response.
      *
-     * @param responseBody The response body, converted to an instance of it's associated Java
-     *                     class.
+     * @param response The response.
      */
-    protected abstract void onResponse(@NonNull final T responseBody);
+    protected abstract void onResponse(@NonNull final Response response);
 
     /**
-     * Callback method for when the HTTP response was not received successfully, whether due to a
-     * network failure, receiving an HTTP error status code, or encountering an unexpected exception
-     * or error during the request creation or response processing phase.
+     * Callback method for when the HTTP response was not received successfully, whether due to
+     * cancellation, a connectivity problem, or a timeout, or receiving an HTTP error status code.
      *
-     * @param error An {@link IOException} if the request failed due to a network failure, an
-     *              {HttpResponseStatusException} if the failure was due to receiving an error code,
-     *              or any {@link Throwable} implementation if one was thrown unexpectedly while
-     *              creating the request or processing the response.
+     * @param error An {@link IOException} if the request failed due to a network failure, or an
+     *              {HttpResponseStatusException} if the failure was due to receiving an error code.
      */
     protected void onFailure(@NonNull final Throwable error) {}
 }
