+++ OpenEdXMobile/src/main/java/org/edx/mobile/http/interceptor/UserAgentInterceptor.java
@@ -1,4 +1,4 @@
-package org.edx.mobile.http;
+package org.edx.mobile.http.interceptor;
 
 import android.support.annotation.NonNull;
