+++ OpenEdXMobile/src/main/java/org/edx/mobile/profiles/UserProfileInteractor.java
@@ -6,7 +6,7 @@ import android.text.TextUtils;
 
 import org.edx.mobile.event.AccountDataLoadedEvent;
 import org.edx.mobile.event.ProfilePhotoUpdatedEvent;
-import org.edx.mobile.http.Callback;
+import org.edx.mobile.http.callback.Callback;
 import org.edx.mobile.logger.Logger;
 import org.edx.mobile.model.api.ProfileModel;
 import org.edx.mobile.module.prefs.UserPrefs;
