+++ OpenEdXMobile/src/main/java/org/edx/mobile/profiles/UserProfileTabsInteractor.java
@@ -3,7 +3,7 @@ package org.edx.mobile.profiles;
 import android.support.annotation.NonNull;
 
 import org.edx.mobile.R;
-import org.edx.mobile.http.Callback;
+import org.edx.mobile.http.callback.Callback;
 import org.edx.mobile.model.Page;
 import org.edx.mobile.user.UserService;
 import org.edx.mobile.util.Config;
