+++ OpenEdXMobile/src/test/java/org/edx/mobile/event/NewVersionAvailableEventTest.java
@@ -7,7 +7,7 @@ import android.support.annotation.Nullable;
 import com.google.gson.internal.bind.util.ISO8601Utils;
 
 import org.edx.mobile.BuildConfig;
-import org.edx.mobile.http.NewVersionBroadcastInterceptor;
+import org.edx.mobile.http.interceptor.NewVersionBroadcastInterceptor;
 import org.edx.mobile.test.BaseTestCase;
 import org.edx.mobile.util.Version;
 import org.junit.AssumptionViolatedException;
@@ -34,8 +34,8 @@ import okhttp3.Response;
 import static org.assertj.core.api.Java6Assertions.assertThat;
 import static org.edx.mobile.http.HttpStatus.ACCEPTED;
 import static org.edx.mobile.http.HttpStatus.UPGRADE_REQUIRED;
-import static org.edx.mobile.http.NewVersionBroadcastInterceptor.HEADER_APP_LATEST_VERSION;
-import static org.edx.mobile.http.NewVersionBroadcastInterceptor.HEADER_APP_VERSION_LAST_SUPPORTED_DATE;
+import static org.edx.mobile.http.interceptor.NewVersionBroadcastInterceptor.HEADER_APP_LATEST_VERSION;
+import static org.edx.mobile.http.interceptor.NewVersionBroadcastInterceptor.HEADER_APP_VERSION_LAST_SUPPORTED_DATE;
 import static org.edx.mobile.test.util.TimeUtilsForTests.DEFAULT_TIME;
 import static org.hamcrest.CoreMatchers.nullValue;
 import static org.junit.Assert.assertEquals;
