+++ OpenEdXMobile/src/test/java/org/edx/mobile/test/http/RetrofitOkhttpTestCase.java
@@ -2,9 +2,9 @@ package org.edx.mobile.test.http;
 
 import android.content.Context;
 
-import org.edx.mobile.http.GzipRequestInterceptor;
+import org.edx.mobile.http.interceptor.GzipRequestInterceptor;
 import org.edx.mobile.http.HttpStatus;
-import org.edx.mobile.http.OauthHeaderRequestInterceptor;
+import org.edx.mobile.http.interceptor.OauthHeaderRequestInterceptor;
 import org.edx.mobile.test.BaseTestCase;
 import org.junit.Ignore;
 import org.robolectric.RuntimeEnvironment;
