+++ OpenEdXMobile/src/test/java/org/edx/mobile/view/CourseBaseActivityTest.java
@@ -8,7 +8,7 @@ import android.view.Gravity;
 import android.widget.ProgressBar;
 
 import org.edx.mobile.R;
-import org.edx.mobile.http.OkHttpUtil;
+import org.edx.mobile.http.util.OkHttpUtil;
 import org.edx.mobile.model.api.EnrolledCoursesResponse;
 import org.edx.mobile.model.course.CourseComponent;
 import org.junit.Test;
