+++ OpenEdXMobile/src/test/java/org/edx/mobile/view/CourseOutlineActivityTest.java
@@ -8,7 +8,7 @@ import android.view.View;
 import android.widget.ListView;
 
 import org.edx.mobile.R;
-import org.edx.mobile.http.OkHttpUtil;
+import org.edx.mobile.http.util.OkHttpUtil;
 import org.edx.mobile.model.api.EnrolledCoursesResponse;
 import org.edx.mobile.model.course.BlockPath;
 import org.edx.mobile.model.course.BlockType;
