+++ OpenEdXMobile/src/test/java/org/edx/mobile/view/CourseUnitNavigationActivityTest.java
@@ -14,7 +14,7 @@ import android.widget.TextView;
 import com.google.gson.JsonObject;
 
 import org.edx.mobile.R;
-import org.edx.mobile.http.OkHttpUtil;
+import org.edx.mobile.http.util.OkHttpUtil;
 import org.edx.mobile.model.Filter;
 import org.edx.mobile.model.api.EnrolledCoursesResponse;
 import org.edx.mobile.model.course.BlockType;
