+++ OpenEdXMobile/src/test/java/org/edx/mobile/view/CourseUnitVideoFragmentTest.java
@@ -14,7 +14,7 @@ import android.view.WindowManager;
 import android.widget.LinearLayout;
 
 import org.edx.mobile.R;
-import org.edx.mobile.http.OkHttpUtil;
+import org.edx.mobile.http.util.OkHttpUtil;
 import org.edx.mobile.model.api.EnrolledCoursesResponse;
 import org.edx.mobile.model.course.CourseComponent;
 import org.edx.mobile.model.course.VideoBlockModel;
